requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
groq
python-dotenv
httpx
